Move CacheSize option to core, OnlineSink to the Builder projects
