+++ summingbird-builder/src/main/scala/com/twitter/summingbird/sink/OnlineSink.scala
@@ -14,7 +14,7 @@
  limitations under the License.
  */
 
-package com.twitter.summingbird.online
+package com.twitter.summingbird.sink
 
 import com.twitter.util.Future
