+++ summingbird-core/src/main/scala/com/twitter/summingbird/option/CacheSize.scala
